Made APIs restful
