+++ microceph/api/types/replication_rbd.go
@@ -1,5 +1,13 @@
 package types
 
+import (
+	"fmt"
+	"net/url"
+	"strings"
+
+	"github.com/canonical/lxd/shared/logger"
+)
+
 // Types for RBD Pool status table.
 type RbdPoolStatusImageBrief struct {
 	Name            string `json:"name" yaml:"name"`
@@ -56,3 +64,134 @@ type RbdPoolBrief struct {
 }
 
 type RbdPoolList []RbdPoolBrief
+
+// ################################## RBD Replication Request ##################################
+// RbdReplicationDirection defines Rbd mirror direction
+type RbdReplicationDirection string
+
+const (
+	RbdReplicationDirectionRXOnly RbdReplicationDirection = "rx-only"
+	RbdReplicationDirectionRXTX   RbdReplicationDirection = "rx-tx"
+)
+
+// RbdResourceType defines request resource type
+type RbdResourceType string
+
+const (
+	RbdResourceDisabled RbdResourceType = "disabled"
+	RbdResourcePool     RbdResourceType = "pool"
+	RbdResourceImage    RbdResourceType = "image"
+)
+
+// RbdReplicationType defines mode of rbd mirroring
+type RbdReplicationType string
+
+const (
+	RbdReplicationDisabled   RbdReplicationType = "disable"
+	RbdReplicationJournaling RbdReplicationType = "journal"
+	RbdReplicationSnapshot   RbdReplicationType = "snapshot"
+)
+
+// RbdReplicationRequest implements ReplicationRequest for RBD replication.
+type RbdReplicationRequest struct {
+	SourcePool  string `json:"source_pool" yaml:"source_pool"`
+	SourceImage string `json:"source_image" yaml:"source_image"`
+	RemoteName  string `json:"remote" yaml:"remote"`
+	// snapshot in d,h,m format
+	Schedule        string                 `json:"schedule" yaml:"schedule"`
+	ReplicationType RbdReplicationType     `json:"replication_type" yaml:"replication_type"`
+	ResourceType    RbdResourceType        `json:"resource_type" yaml:"resource_type"`
+	RequestType     ReplicationRequestType `json:"request_type" yaml:"request_type"`
+	IsForceOp       bool                   `json:"force" yaml:"force"`
+	SkipAutoEnable  bool                   `json:"skipAutoEnable" yaml:"skipAutoEnable"`
+}
+
+// GetWorkloadType provides the workload name for replication request
+func (req RbdReplicationRequest) GetWorkloadType() CephWorkloadType {
+	return RbdWorkload
+}
+
+// GetAPIObjectId provides the API object id i.e. /replication/rbd/<object-id>
+func (req RbdReplicationRequest) GetAPIObjectId() string {
+	// If both Pool and Image values are present encode for query.
+	if len(req.SourceImage) != 0 && len(req.SourcePool) != 0 {
+		resource := url.QueryEscape(fmt.Sprintf("%s/%s", req.SourcePool, req.SourceImage))
+		logger.Debugf("REPAPI: Resource: %s", resource)
+		return resource
+	}
+
+	return req.SourcePool
+}
+
+// SetAPIObjectId provides the API object id i.e. /replication/rbd/<object-id>
+func (req *RbdReplicationRequest) SetAPIObjectId(id string) error {
+	// unescape object string
+	object, err := url.PathUnescape(id)
+	if err != nil {
+		return err
+	}
+
+	frags := strings.Split(string(object), "/")
+	if len(frags) > 1 {
+		req.SourcePool = frags[0]
+		req.SourceImage = frags[1]
+	} else {
+		req.SourcePool = object
+	}
+
+	return nil
+}
+
+// GetAPIRequestType provides the REST method for the request
+func (req RbdReplicationRequest) GetAPIRequestType() string {
+	frags := strings.Split(string(req.RequestType), "-")
+	logger.Debugf("REPAPI: API frags: %v", frags)
+	if len(frags) == 0 {
+		return ""
+	}
+
+	return frags[0]
+}
+
+// GetWorkloadRequestType provides the event used as the FSM trigger.
+func (req RbdReplicationRequest) GetWorkloadRequestType() string {
+	frags := strings.Split(string(req.RequestType), "-")
+	logger.Debugf("REPAPI: Workload frags: %v", frags)
+	if len(frags) < 2 {
+		return ""
+	}
+
+	return frags[1]
+}
+
+// ################### Helpers ############################
+// GetRbdResourceType gets the resource type of the said request
+func GetRbdResourceType(poolName string, imageName string) RbdResourceType {
+	if len(poolName) != 0 && len(imageName) != 0 {
+		return RbdResourceImage
+	} else {
+		return RbdResourcePool
+	}
+}
+
+func GetPoolAndImageFromResource(resource string) (string, string, error) {
+	var pool string
+	var image string
+	resourceFrags := strings.Split(resource, "/")
+	if len(resourceFrags) < 1 || len(resourceFrags) > 2 {
+		return "", "", fmt.Errorf("check resource name %s, should be in $pool/$image format", resource)
+	}
+
+	// If only pool name is provided.
+	if len(resourceFrags) == 1 {
+		pool = resourceFrags[0]
+		image = ""
+	} else
+	// if both pool and image names are provided.
+	if len(resourceFrags) == 2 {
+		pool = resourceFrags[0]
+		image = resourceFrags[1]
+	}
+
+	return pool, image, nil
+}
